% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_api_key.r
\name{algo_api_key}
\alias{algo_api_key}
\title{Get or set ALGORITHMIA_API_KEY value}
\usage{
algo_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new Algorithmia API key for the current environment?}
}
\value{
atomic character vector containing the Algorithmia API key
}
\description{
The API wrapper functions in this package all rely on a Algorithmia API
key residing in the environment variable \code{ALGORITHMIA_API_KEY}. The
easiest way to accomplish this is to set it in the `.Renviron` file in your
home directory.
}

