% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_call}
\alias{algo_call}
\title{Call an Algorithmia algorithm}
\usage{
algo_call(algo_obj, owner, name, version = NULL)
}
\arguments{
\item{owner}{owner}

\item{name}{name}

\item{version}{version (optional)}
}
\description{
Call an Algorithmia algorithm
}
\examples{
library(magrittr)
algo_client() \%>\%
  algo_call("demo", "Hello", "0.1.1") \%>\%
  algo_pipe("there", "text")
}

