% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_client}
\alias{algo_client}
\title{Initialize a new Algorithmia API client call}
\usage{
algo_client(api_key = algo_api_key())
}
\arguments{
\item{api_key}{Algorithmia API key}
}
\description{
Initialize a new Algorithmia API client call
}
\examples{
library(magrittr)
algo_client() \%>\%
  algo_call("demo", "Hello", "0.1.1") \%>\%
  algo_pipe("there", "text")
}

