% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_dir_list}
\alias{algo_dir_list}
\title{List a directory accessible by the Algorithmia service}
\usage{
algo_dir_list(algo_obj, dir_spec)
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{dir_spec}{URI (See \url{http://docs.algorithmia.com/#data-uri})}
}
\description{
List a directory accessible by the Algorithmia service
}
\references{
\url{http://docs.algorithmia.com/}
}
