% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_download_file}
\alias{algo_download_file}
\title{Read a file from a URI accessible by the Algorithmia service & write to disk}
\usage{
algo_download_file(algo_obj, fil_spec, local_fil, overwrite = TRUE)
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{fil_spec}{URI}

\item{overwrite}{overwrite file if it exists?}

\item{local_file}{local file name to write to}
}
\description{
Useful if the file is larger than memory.
}
\references{
\url{http://docs.algorithmia.com/}
}

