% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_file_exists}
\alias{algo_file_exists}
\title{Test if a file exists}
\usage{
algo_file_exists(algo_obj, fil_spec)
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{fil_spec}{URI}
}
\description{
Test if a file exists
}

