% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_options}
\alias{algo_options}
\title{Set options for an algorithm call}
\usage{
algo_options(algo_obj, timeout = 300, stdout = FALSE, output = NULL)
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{timeout}{timeout}

\item{stdout}{stdout}

\item{output}{output}
}
\description{
Set options for an algorithm call
}
\examples{
library(magrittr)
algo_client() \%>\%
  algo_call("demo", "Hello", "0.1.1") \%>\%
  algo_options(30) \%>\%
  algo_pipe("there", "text")
}

