% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_pipe}
\alias{algo_pipe}
\title{Execute an Algorithmia algorithm call}
\usage{
algo_pipe(algo_obj, input, content_type = c("text", "json"))
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{input}{data to feed to the API call}

\item{content_type}{specify encoding}
}
\description{
Execute an Algorithmia algorithm call
}
\examples{
library(magrittr)
algo_client() \%>\%
  algo_call("demo", "Hello", "0.1.1") \%>\%
  algo_options(30) \%>\%
  algo_pipe("there", "text")
}

