% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_read_file}
\alias{algo_read_file}
\title{Read a file from a URI accessible by the Algorithmia service}
\usage{
algo_read_file(algo_obj, fil_spec, fmt = c("text", "parsed", "raw"),
  encoding = NULL)
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{fil_spec}{URI}

\item{fmt}{result object format}

\item{encoding}{content encoding}
}
\description{
Read a file from a URI accessible by the Algorithmia service
}

