% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithmia-package.r
\docType{package}
\name{algorithmia}
\alias{algorithmia}
\alias{algorithmia-package}
\title{An R interface to the Algorithmia API}
\description{
An R interface to the Algorithmia API
}
\author{
Bob Rudis (@hrbrmstr)
}

