% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_dir_exists}
\alias{algo_dir_exists}
\title{Test if a directory exists}
\usage{
algo_dir_exists(algo_obj, dir_spec)
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{dir_spec}{URI}
}
\description{
Test if a directory exists
}

