% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.r
\name{algo_read_file}
\alias{algo_read_file}
\title{Read a file from a URI accessible by the Algorithmia service}
\usage{
algo_read_file(algo_obj, fil_spec, fmt = c("text", "parsed", "raw"),
  encoding = NULL)
}
\arguments{
\item{algo_obj}{an \code{algorithmia} object}

\item{fil_spec}{URI (See \url{http://docs.algorithmia.com/#data-uri})}

\item{fmt}{result object format; "\code{text}" to retrieve the plain text of the
response; "\code{parsed}" to parse the response and return an R object; "\code{raw}"
to retrieve the binary response (useful for images)}

\item{encoding}{content encoding}
}
\description{
Read a file from a URI accessible by the Algorithmia service
}
\references{
\url{http://docs.algorithmia.com/}
}
